+++ kogito-drools/drools-core/src/main/java/org/drools/core/spi/FactHandleFactory.java
@@ -37,7 +37,7 @@ public interface FactHandleFactory {
                                      InternalWorkingMemory workingMemory,
                                      WorkingMemoryEntryPoint wmEntryPoint );
     
-    InternalFactHandle newFactHandle(int id,
+    InternalFactHandle newFactHandle(long id,
                                      Object object,
                                      long recency,
                                      ObjectTypeConf conf,
@@ -59,20 +59,20 @@ public interface FactHandleFactory {
      */
     FactHandleFactory newInstance();
     
-    FactHandleFactory newInstance(int id, long counter);
+    FactHandleFactory newInstance(long id, long counter);
 
     Class<?> getFactHandleType();
 
-    int getId();
+    long getId();
 
     long getRecency();
 
-    int getNextId();
+    long getNextId();
 
     long getNextRecency();
     
-    void clear(int id, long counter);
+    void clear(long id, long counter);
 
-    void doRecycleIds( Collection<Integer> usedIds );
+    void doRecycleIds( Collection<Long> usedIds );
     void stopRecycleIds();
 }
